+++ src/main/api.ts
@@ -24,6 +24,7 @@ export const ApiRouteNames = [
 
   'window:setTitle',
   'window:destroy',
+  'window:setZoom',
 
   'clipboard:writeWournal',
   'clipboard:readText',
@@ -58,6 +59,7 @@ export interface ElectronApi extends ApiSpec<ApiRouteName> {
 
   'window:setTitle': (title: string) => Promise<void>;
   'window:destroy': () => Promise<void>;
+  'window:setZoom': (zoom: number) => Promise<void>;
 
   'clipboard:writeWournal': (data: any) => Promise<void>;
   'clipboard:readText': () => Promise<string | false>;